+++ examples/resources/cloudflare_authenticated_origin_pulls/import.sh
@@ -1,8 +1,8 @@
-# Authenticated Origin Pull configuration
-$ terraform import cloudflare_authenticated_origin_pulls_certificate.my_aop <zone_id>//
+# global
+$ terraform import cloudflare_authenticated_origin_pulls.example <zone_id>
 
-# Per-Zone Authenticated Origin Pull configuration
-$ terraform import cloudflare_authenticated_origin_pulls_certificate.my_per_zone_aop <zone_id>/<certificate_id>/
+# per zone
+$ terraform import cloudflare_authenticated_origin_pulls.example <zone_id>/<certificate_id>
 
-# Per-Hostname Authenticated Origin Pull configuration
-$ terraform import cloudflare_authenticated_origin_pulls_certificate.my_per_hostname_aop <zone_id>/<certificate_id>/<hostname>
+# per hostname
+$ terraform import cloudflare_authenticated_origin_pulls.example <zone_id>/<certificate_id>/<hostname>