+++ internal/framework/service/rulesets/schema.go
@@ -154,6 +154,11 @@ func (r *RulesetResource) Schema(ctx context.Context, req resource.SchemaRequest
 							MarkdownDescription: "List of parameters that configure the behavior of the ruleset rule action.",
 							NestedObject: schema.NestedBlockObject{
 								Attributes: map[string]schema.Attribute{
+									"additional_cacheable_ports": schema.SetAttribute{
+										ElementType:         types.Int64Type,
+										Optional:            true,
+										MarkdownDescription: "Specifies uncommon ports to allow cacheable assets to be served from.",
+									},
 									"automatic_https_rewrites": schema.BoolAttribute{
 										Optional:            true,
 										MarkdownDescription: "Turn on or off Cloudflare Automatic HTTPS rewrites.",
@@ -465,13 +470,13 @@ func (r *RulesetResource) Schema(ctx context.Context, req resource.SchemaRequest
 											Attributes: map[string]schema.Attribute{
 												"mode": schema.StringAttribute{
 													Required:            true,
-													Validators:          []validator.String{stringvalidator.OneOf("override_origin", "respect_origin")},
-													MarkdownDescription: "Mode of the edge TTL.",
+													Validators:          []validator.String{stringvalidator.OneOf("override_origin", "respect_origin", "bypass_by_default")},
+													MarkdownDescription: fmt.Sprintf("Mode of the edge TTL. %s", utils.RenderAvailableDocumentationValuesStringSlice([]string{"override_origin", "respect_origin", "bypass_by_default"})),
 												},
 												"default": schema.Int64Attribute{
 													Optional:            true,
 													Validators:          []validator.Int64{int64validator.AtLeast(1)},
-													MarkdownDescription: "Default edge TTL",
+													MarkdownDescription: "Default edge TTL.",
 												},
 											},
 											Validators: []validator.Object{EdgeTTLValidator{}},
@@ -917,7 +922,7 @@ func (r *RulesetResource) Schema(ctx context.Context, req resource.SchemaRequest
 										MarkdownDescription: "Criteria for counting HTTP requests to trigger the Rate Limiting action. Uses the Firewall Rules expression language based on Wireshark display filters. Refer to the [Firewall Rules language](https://developers.cloudflare.com/firewall/cf-firewall-language) documentation for all available fields, operators, and functions.",
 									},
 									"requests_to_origin": schema.BoolAttribute{
-										Optional:            true,
+										Required:            true,
 										MarkdownDescription: "Whether to include requests to origin within the Rate Limiting count.",
 									},
 								},
@@ -945,7 +950,7 @@ func (r *RulesetResource) Schema(ctx context.Context, req resource.SchemaRequest
 							},
 						},
 						"logging": schema.ListNestedBlock{
-							MarkdownDescription: "List parameters to configure how the rule generates logs.",
+							MarkdownDescription: "List parameters to configure how the rule generates logs. Only valid for skip action.",
 							NestedObject: schema.NestedBlockObject{
 								Attributes: map[string]schema.Attribute{
 									"enabled": schema.BoolAttribute{