+++ internal/sdkv2provider/data_source_account_roles.go
@@ -5,10 +5,11 @@ import (
 	"fmt"
 
 	"github.com/cloudflare/cloudflare-go"
-	"github.com/cloudflare/terraform-provider-cloudflare/internal/consts"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+	"github.com/cloudflare/terraform-provider-cloudflare/internal/consts"
 )
 
 func dataSourceCloudflareAccountRolesRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
@@ -16,7 +17,7 @@ func dataSourceCloudflareAccountRolesRead(ctx context.Context, d *schema.Resourc
 	accountID := d.Get(consts.AccountIDSchemaKey).(string)
 
 	tflog.Debug(ctx, fmt.Sprintf("Reading Account Roles"))
-	roles, err := client.AccountRoles(ctx, accountID)
+	roles, err := client.ListAccountRoles(ctx, cloudflare.AccountIdentifier(accountID), cloudflare.ListAccountRolesParams{})
 	if err != nil {
 		return diag.FromErr(fmt.Errorf("error listing Account Roles: %w", err))
 	}