+++ internal/sdkv2provider/resource_cloudflare_access_identity_provider.go
@@ -62,7 +62,10 @@ func resourceCloudflareAccessIdentityProviderRead(ctx context.Context, d *schema
 		return diag.FromErr(fmt.Errorf("error setting Access Identity Provider configuration: %w", configErr))
 	}
 
-	scimConfig := convertAccessIDPScimConfigStructToSchema(accessIdentityProvider.ScimConfig)
+	// We need to get the current secret and set that in the state as it is only
+	// returned initially when the resource was created. The value read from the
+	// server will always be redacted.
+	scimConfig := convertAccessIDPScimConfigStructToSchema(d.Get("scim_config.0.secret").(string), accessIdentityProvider.ScimConfig)
 	if scimConfigErr := d.Set("scim_config", scimConfig); scimConfigErr != nil {
 		return diag.FromErr(fmt.Errorf("error setting Access Identity Provider scim configuration: %w", scimConfigErr))
 	}
@@ -96,6 +99,11 @@ func resourceCloudflareAccessIdentityProviderCreate(ctx context.Context, d *sche
 	}
 
 	d.SetId(accessIdentityProvider.ID)
+
+	scimConfig := convertAccessIDPScimConfigStructToSchema(accessIdentityProvider.ScimConfig.Secret, accessIdentityProvider.ScimConfig)
+	if scimConfigErr := d.Set("scim_config", scimConfig); scimConfigErr != nil {
+		return diag.FromErr(fmt.Errorf("error setting Access Identity Provider scim configuration: %w", scimConfigErr))
+	}
 
 	return resourceCloudflareAccessIdentityProviderRead(ctx, d, meta)
 }
@@ -133,6 +141,16 @@ func resourceCloudflareAccessIdentityProviderUpdate(ctx context.Context, d *sche
 
 	if accessIdentityProvider.ID == "" {
 		return diag.FromErr(fmt.Errorf("failed to find Access Identity Provider ID in update response; resource was empty"))
+	}
+
+	scimSecret := d.Get("scim_config.0.secret").(string)
+	if accessIdentityProvider.ScimConfig.Secret != "" && !strings.Contains(accessIdentityProvider.ScimConfig.Secret, "*") {
+		scimSecret = accessIdentityProvider.ScimConfig.Secret
+	}
+
+	scimConfig := convertAccessIDPScimConfigStructToSchema(scimSecret, accessIdentityProvider.ScimConfig)
+	if scimConfigErr := d.Set("scim_config", scimConfig); scimConfigErr != nil {
+		return diag.FromErr(fmt.Errorf("error setting Access Identity Provider scim configuration: %w", scimConfigErr))
 	}
 
 	return resourceCloudflareAccessIdentityProviderRead(ctx, d, meta)
@@ -215,10 +233,13 @@ func convertSchemaToStruct(d *schema.ResourceData) (cloudflare.AccessIdentityPro
 		IDPConfig.ClientSecret = d.Get("config.0.client_secret").(string)
 		IDPConfig.DirectoryID = d.Get("config.0.directory_id").(string)
 		IDPConfig.EmailAttributeName = d.Get("config.0.email_attribute_name").(string)
+		IDPConfig.EmailClaimName = d.Get("config.0.email_claim_name").(string)
 		IDPConfig.IdpPublicCert = d.Get("config.0.idp_public_cert").(string)
 		IDPConfig.IssuerURL = d.Get("config.0.issuer_url").(string)
 		IDPConfig.OktaAccount = d.Get("config.0.okta_account").(string)
+		IDPConfig.OktaAuthorizationServerID = d.Get("config.0.authorization_server_id").(string)
 		IDPConfig.OneloginAccount = d.Get("config.0.onelogin_account").(string)
+		IDPConfig.PingEnvID = d.Get("config.0.ping_env_id").(string)
 		IDPConfig.RedirectURL = d.Get("config.0.redirect_url").(string)
 		IDPConfig.SignRequest = d.Get("config.0.sign_request").(bool)
 		IDPConfig.SsoTargetURL = d.Get("config.0.sso_target_url").(string)
@@ -256,6 +277,7 @@ func convertAccessIDPConfigStructToSchema(options cloudflare.AccessIdentityProvi
 		"apps_domain":                options.AppsDomain,
 		"attributes":                 attributes,
 		"auth_url":                   options.AuthURL,
+		"authorization_server_id":    options.OktaAuthorizationServerID,
 		"centrify_account":           options.CentrifyAccount,
 		"centrify_app_id":            options.CentrifyAppID,
 		"certs_url":                  options.CertsURL,
@@ -265,10 +287,12 @@ func convertAccessIDPConfigStructToSchema(options cloudflare.AccessIdentityProvi
 		"scopes":                     options.Scopes,
 		"directory_id":               options.DirectoryID,
 		"email_attribute_name":       options.EmailAttributeName,
+		"email_claim_name":           options.EmailClaimName,
 		"idp_public_cert":            options.IdpPublicCert,
 		"issuer_url":                 options.IssuerURL,
 		"okta_account":               options.OktaAccount,
 		"onelogin_account":           options.OneloginAccount,
+		"ping_env_id":                options.PingEnvID,
 		"redirect_url":               options.RedirectURL,
 		"sign_request":               options.SignRequest,
 		"sso_target_url":             options.SsoTargetURL,
@@ -281,10 +305,10 @@ func convertAccessIDPConfigStructToSchema(options cloudflare.AccessIdentityProvi
 	return []interface{}{m}
 }
 
-func convertAccessIDPScimConfigStructToSchema(options cloudflare.AccessIdentityProviderScimConfiguration) []interface{} {
+func convertAccessIDPScimConfigStructToSchema(secret string, options cloudflare.AccessIdentityProviderScimConfiguration) []interface{} {
 	m := map[string]interface{}{
 		"enabled":                  options.Enabled,
-		"secret":                   options.Secret,
+		"secret":                   secret,
 		"user_deprovision":         options.UserDeprovision,
 		"seat_deprovision":         options.SeatDeprovision,
 		"group_member_deprovision": options.GroupMemberDeprovision,